Fitting linear models using farray

Fitting linear models using farray

farray_lm(
  formula,
  data,
  fitted = FALSE,
  weights = NULL,
  offset = NULL,
  contrasts = NULL,
  na.action = getOption("na.action"),
  qr.tol = 1e-07,
  ...
)

Arguments

formula
an object of class stats::formula(). For variable names to be used, see 'Details'

data
a farray object

fitted
whether to calculate fitted data and residuals. This may take time and memory if data is large

weights, offset, contrasts, na.action
see stats::lm()

qr.tol
the tolerance for detecting linear dependencies in the partitions of data; see qr()

...
passed to chunk_map()

Value

An object of class c("farray_lm", "lm") or for multiple responses of class c("farray_lm", "mlm").

Details

The array will be reshaped to a matrix first before fitting the linear models. A \(100 x 20 x 5\) array will be reshaped to a \(2000 x 5\) file matrix. The variables are the partitions of the array. If dimnames are set for the last margin index, then those will be used as variable names, otherwise farray_lm automatically assign \("V1", "V2", "V3", ...\) as each partition names.

Examples

library(farray)
arr <- array(rnorm(72), c(6,3,4))
arr[1,1,1] <- NA        # Allow NA to be treated
offset = rnorm(18)    # offset and weights are optional
weights = runif(18)

formula <- V1 ~ .-V2-1 + (V2 > 0)

data <- as.farray(arr)
object <- farray_lm(formula, data, weights = weights, offset = offset)

# Compare to stats::lm
dim(arr) <- c(18, 4)
lm_data <- as.data.frame(arr)
flm <- lm(formula, lm_data, weights = weights, offset = offset)

cbind(coef(object), coef(flm))

#>                   [,1]       [,2]
#> V3           1.0832081  1.0832081
#> V4          -0.5884334 -0.5884334
#> V2 > 0FALSE -0.5648055 -0.5648055
#> V2 > 0TRUE  -0.5706421 -0.5706421

cbind(resid(object), resid(flm))

#>            [,1]         [,2]
#> 2   2.447436087  2.447436087
#> 3   0.005134941  0.005134941
#> 4   1.476781486  1.476781486
#> 5  -0.442949248 -0.442949248
#> 6  -1.158822413 -1.158822413
#> 7  -0.719018900 -0.719018900
#> 8   3.423522142  3.423522142
#> 9  -0.677011081 -0.677011081
#> 10 -0.995005208 -0.995005208
#> 11 -1.378474872 -1.378474872
#> 12 -3.461595746 -3.461595746
#> 13  1.837165104  1.837165104
#> 14 -0.342684597 -0.342684597
#> 15 -0.547523398 -0.547523398
#> 16  0.665466858  0.665466858
#> 17 -0.329242384 -0.329242384
#> 18  0.807315088  0.807315088

cbind(fitted(object), fitted(flm))

#>           [,1]        [,2]
#> 2  -1.38012821 -1.38012821
#> 3   0.06489991  0.06489991
#> 4  -2.11590481 -2.11590481
#> 5   0.39298435  0.39298435
#> 6   0.90733897  0.90733897
#> 7   1.16381602  1.16381602
#> 8  -0.66810457 -0.66810457
#> 9   0.72354246  0.72354246
#> 10  1.57271428  1.57271428
#> 11  1.49666975  1.49666975
#> 12  1.54987526  1.54987526
#> 13 -0.97507862 -0.97507862
#> 14  0.09944786  0.09944786
#> 15  0.34143620  0.34143620
#> 16 -0.64628927 -0.64628927
#> 17  0.35880314  0.35880314
#> 18 -0.25748755 -0.25748755

summary(object)

#> 
#> Call:
#> farray_lm(formula = formula, data = data, weights = weights, 
#>     offset = offset)
#> 
#> Weighted Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -2.1572 -0.4941 -0.1571  0.6473  2.9807 
#> 
#> Coefficients:
#>      Estimate Std. Error t value Pr(>|t|)  
#> [1,]   1.0832     0.5620   1.927   0.0761 .
#> [2,]  -0.5884     0.5665  -1.039   0.3179  
#> [3,]  -0.5648     0.6267  -0.901   0.3839  
#> [4,]  -0.5706     0.6693  -0.853   0.4093  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 1.206 on 13 degrees of freedom
#> Multiple R-squared:  0.2645,	Adjusted R-squared:  0.03816 
#> F-statistic: 1.169 on 4 and 13 DF,  p-value: 0.3692
#> 

summary(flm)

#> 
#> Call:
#> lm(formula = formula, data = lm_data, weights = weights, offset = offset)
#> 
#> Weighted Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -2.1572 -0.4941 -0.1571  0.6473  2.9807 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)  
#> V3            1.0832     0.5620   1.927   0.0761 .
#> V4           -0.5884     0.5665  -1.039   0.3179  
#> V2 > 0FALSE  -0.5648     0.6267  -0.901   0.3839  
#> V2 > 0TRUE   -0.5706     0.6693  -0.853   0.4093  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 1.206 on 13 degrees of freedom
#>   (1 observation deleted due to missingness)
#> Multiple R-squared:  0.2645,	Adjusted R-squared:  0.03816 
#> F-statistic: 1.169 on 4 and 13 DF,  p-value: 0.3692
#>